requirements.txt
nose==1.3.7
m3-cdecimal==2.3